requirements.txt
streamlit==1.32.0
spacy==3.7.4
google-generativeai==0.4.1
python-dotenv==1.0.1
textstat==0.7.3
jsonschema==4.21.1
PyPDF2==3.0.1
fpdf==1.7.2
python-docx==1.1.0
pdfplumber==0.11.0
sentence-transformers>=3.0.0
scikit-learn==1.3.2
nltk==3.8.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
plotly>=5.19.0
networkx>=3.2.1
pandas>=2.2.0
altair>=5.0.0

# Dev Dependencies
pytest
mypy
black
